chore: cleanup old code and move some files
