+++ cmd/relationtuple/create.go
@@ -7,12 +7,13 @@ import (
 	"io"
 	"os"
 
+	"github.com/ory/keto/relationtuple"
+
 	"github.com/spf13/cobra"
 
 	"github.com/ory/x/cmdx"
 
 	"github.com/ory/keto/cmd/client"
-	"github.com/ory/keto/models"
 )
 
 func newCreateCmd() *cobra.Command {
@@ -36,16 +37,16 @@ func newCreateCmd() *cobra.Command {
 				}
 			}
 
-			var r models.InternalRelationTuple
+			var r relationtuple.InternalRelationTuple
 			err = json.NewDecoder(f).Decode(&r)
 			if err != nil {
 				_, _ = fmt.Fprintf(cmd.ErrOrStderr(), "Could not decode: %s\n", err)
 				return cmdx.FailSilently(cmd)
 			}
 
-			cl := models.NewRelationTupleServiceClient(conn)
+			cl := relationtuple.NewRelationTupleServiceClient(conn)
 
-			_, err = cl.WriteRelationTuple(context.Background(), &models.WriteRelationTupleRequest{Tuple: (&models.RelationTuple{}).FromInternal(&r)})
+			_, err = cl.WriteRelationTuple(context.Background(), &relationtuple.WriteRelationTupleRequest{Tuple: (&relationtuple.RelationTuple{}).FromInternal(&r)})
 			if err != nil {
 				_, _ = fmt.Fprintf(cmd.ErrOrStderr(), "Error doing the request: %s\n", err)
 				return cmdx.FailSilently(cmd)
