+++ cmd/serve.go
@@ -32,7 +32,6 @@ import (
 	"github.com/ory/graceful"
 
 	"github.com/ory/keto/driver"
-	"github.com/ory/keto/models"
 	"github.com/ory/keto/relationtuple"
 
 	"github.com/ory/x/viperx"
@@ -70,7 +69,7 @@ on configuration options, open the configuration documentation:
 
 			s := grpc.NewServer()
 			relS := relationtuple.NewServer(reg)
-			models.RegisterRelationTupleServiceServer(s, relS)
+			relationtuple.RegisterRelationTupleServiceServer(s, relS)
 			fmt.Println("going to serve GRPC on", lis.Addr().String())
 			if err := s.Serve(lis); err != nil {
 				fmt.Fprintf(cmd.ErrOrStderr(), "%+v\n", err)
